==> 01 - Data import, decode, join and save.ipynb <==
# 01 - Data importing, decoding, joining, and saving.
___

Step 1! Let's get our hands on the data and select for only collisions in South Yorkshire, as that's what the project focuses on. We'll decode the data to make it human readable, so that we can do some exploratory data analysis to get familiar with the data and have a bit more direction in training our model. We'll join the Casualty and Collision datasets into one big dataframe that we wil export to a .csv to read into our code for Step 2: Data cleaning.

import pandas as pd
import sqlalchemy

Here i'm reading in the information we need to convert the numbers in the downloaded data tables into their human readable equivalent. This is stored in the 'stats19_schema.csv' document

schema = pd.read_csv('stats19_schema.csv')
coll_schema = schema[schema.table=='Accident']
cas_schema = schema[schema.table=='Casualty']

Now time to download the data. I'm only downloading the last 5 years data because this is nice and fast for me to do (gov.uk website has compiled them for us!). This cell will take a little bit of time to run whilst it downloads everything.

coll_url = r"https://data.dft.gov.uk/road-accidents-safety-data/dft-road-casualty-statistics-collision-last-5-years.csv"
cas_url = r"https://data.dft.gov.uk/road-accidents-safety-data/dft-road-casualty-statistics-casualty-last-5-years.csv"

coll_df = pd.read_csv(coll_url, low_memory = False, na_values=[-1, '-1', 'Data missing or out of range', 'Unknown'])
cas_df = pd.read_csv(cas_url, low_memory = False, na_values=[-1, '-1', 'Data missing or out of range', 'Unknown'])

Below is a neat little piece of code that i'm proud of! In just one line of code we're converting all of the data to their corresponding strings. 😎

coll_df = pd.concat([coll_df[name].map(coll_schema[coll_schema.variable == name].set_index('code').label).copy() if name in coll_schema.variable.unique() else coll_df[name] for name in coll_df.columns], axis=1)

cas_df = pd.concat([
    cas_df[name].map(cas_schema[cas_schema.variable == name].set_index('code').label).copy() if (name in cas_schema.variable.unique() and name != 'age_of_casualty') else cas_df[name] for name in cas_df.columns], axis=1)

cas_df.head()

coll_df.head()

___

# 01.1 - Creating tables in a Postgres database

Here, we will use SQLAlchemy and Pandas to move the dataframes into a local Postgres database. For this ML project we will focus on South Yorkshire data, but we can use SQL to poke around in the UK-wide dataset with less issues due to the dataset size (and to show off our SQL chops!).

from sqlalchemy import create_engine, inspect
%load_ext sql

engine = create_engine('postgresql://postgres:postgres@localhost:5432/STATS19_casualty_collision')
try:
  %sql DROP table casualty;
  %sql DROP table collision;
except:
  pass

inspection = inspect(engine)
print(f'Current tables in the database are:\n {inspection.get_table_names()}')

cas_df.to_sql(con=engine, name='casualty', if_exists='replace', index=False)
coll_df.to_sql(con=engine, name='collision', if_exists='replace', index=False)

inspection = inspect(engine)
print(f'Now the tables in the database are:\n {inspection.get_table_names()}')

print(pd.read_sql("""
                  SELECT COUNT(*)
                  FROM casualty;
                  """, engine))

print(pd.read_sql("""
                  SELECT COUNT(*)
                  FROM collision;
                  """, engine))

All done. We'll query our new database later. For now, let's select only the South Yorkshire data and combine the Casualty and Collision dataframes in a one (collision) to many (casualty) join.

___

Now i'm selecting for only the data corresponding to South Yorkshire. Down the line if we want to train on all of the UK data we can just remove this line of code, but beware that we will have tailored our features towards the South Yorkshire data so it might not generalise so well.

coll_df = coll_df[coll_df.police_force == 'South Yorkshire']

We can sum the number of casualties (from the number_of_casualties column) to validate the expected shape of the joined dataframe. 

print(f'There are {coll_df['number_of_casualties'].sum()} total casualties in South Yorkshire over the last 5 years.')

Now we'll merge the two tables of data. Using a left join we drop any casualty data that doeesn't correspond to an accident in South Yorkshire.

joined = coll_df.merge(cas_df, on='accident_index', how='left')

And we will validate the one-to-many join has proceded as expected.

try:
  assert joined.shape[0] == coll_df['number_of_casualties'].sum()
  print(f'The size of the joined dataframe is as expected with {joined.shape[0]} rows.')
except:
  print(f'The dataframe shape is not as expected. There are now {joined.shape[0]} rows.')

And export it to a .csv for us to read in later on.

joined.to_csv('dft_statistics_collision_and_casualty_last_5_years.csv', index=False)
