==> 03 - Preliminary data cleaning.ipynb <==
# 03 - Preliminary data cleaning
____

Step 3! We have our pristine dataset split nicely into our training (and validation) and testing data. 

It is important to remember that any steps / transformations we apply to our training data must also be applied to the test data to avoid any errors. If we're going to drop columns, lets build up a list of cols_to_drop so we can apply this nice and easily.


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt 

pd.options.mode.chained_assignment = None 

train = pd.read_csv('training_data.csv', low_memory = False, na_values=[-1, '-1', 'Data missing or out of range', 'Unknown'])
test = pd.read_csv('testing_data.csv', low_memory = False, na_values=[-1, '-1', 'Data missing or out of range', 'Unknown'])


## Let's have a look at our data types.

train.info()

Firstly, we have a lot of features. Feature selection will be very important to make sure that our model doesn't overfit. There are also a lot of superfluous features that can easily be dropped at our feature selection stage:
  * The geographical location columns 
  * The duplicate accident_year and accident_reference columns
  * Whether or not a police officer attended *after* the collision occured won't have affected the seriousness
  * vehicle and casualty reference isn't useful
  * accident_severity should be dropped as it is highly correlated to the casualty_severity, and will just be a proxy for the model to overfit to
  * same with the enhanced severity

It also looks like we have a tonne of missing data, so first let's make sure that we drop features that have so much missing they won't really assist in our predictions.


### Let's drop columns where we're missing a lot of data.

We have a lot of NaN values, but also some more disguised missing data. Some entries are labelled with -1 (both text and numeric present), 'Data missing or out of range', and 'Unknown'. Missing greater than a third of data is going to limit our predictive power so let's remove these columns from the off. We've specified the alternative values for the NaN values in our read_csv() function above, so these will be counted with typical NaN based methods.

We will evaluate the missing data in the training data set, and apply any column dropping to both datasets. No peeking at the testing data!


columns_to_drop = []

columns_to_drop.extend(train.columns[(train.isna().sum() / train.shape[0]) * 100 > 5].to_list())


print(columns_to_drop)

37% of local authority district data missing, but we have the full dataset for ONS district. Let's use this to infill the local authority district and then drop the ONS district column.

for data in [train, test]:
  districts = {'E08000019': 'Sheffield', 'E08000017': 'Doncaster', 'E08000018': 'Rotherham', 'E08000016': 'Barnsley'}

  for ind,row in enumerate(data.local_authority_district):
    if row is np.nan:
      district = districts[data.local_authority_ons_district.iloc[ind]]
      data.local_authority_district.iloc[ind] = district

columns_to_drop.append('local_authority_ons_district')
columns_to_drop.remove('local_authority_district')

#### Now we're ready to drop the columns with >5% of the data missing. We'll apply this to both the test and train data, and output the saved data.

[data.drop(columns=columns_to_drop, inplace=True) for data in [train, test]]

#### Which columns still have missing data?

cols_missing_data = train.loc[:, train.isnull().any()].isna().sum().sort_values(ascending=False)
print(cols_missing_data)
missing_numeric = train[cols_missing_data.index].describe().columns
missing_categorical = train.describe(exclude=[np.number]).columns

The missing data for these features represents quite a small proportion of the dataset, so let's just infill with the most frequent value for categories, or the mean for numerics. We'll apply this to all columns in the dataframe incase our testing data has missing data in other features.

[[data[col].fillna(data[col].mean(), inplace=True) for col in (data.describe().columns)] for data in [train, test]]

for data in [train, test]:
  categorical_cols = data.describe(exclude=[np.number]).columns
  data[categorical_cols] = data[categorical_cols].apply(lambda x: x.fillna(x.value_counts().index[0]))

train.info()

Let's get rid of the duplicated columns from merging the data.

for data in [train, test]:
  data.rename(columns={'accident_year_x': 'accident_year', 'accident_reference_x': 'accident_reference'}, inplace=True)
  data.drop(labels=['accident_year_y', 'accident_reference_y', 'police_force'], axis=1, inplace=True)

train.to_csv('cleaned_training_data.csv', index=False)
test.to_csv('cleaned_testing_data.csv', index=False)
